+++ src/lib/api/__mocks__/index.ts
@@ -1,2 +1,2 @@
-export { mockProviderAvailabilityResponse } from './providerAvailability';
-export { mockProvidersResponse } from './providers';
+export { mockProviderAvailabilityResponse } from './providerAvailability.mock';
+export { mockProvidersResponse } from './providers.mock';
