+++ tsconfig.json
@@ -23,5 +23,5 @@
   },
   "files": ["globals.d.ts"],
   "include": ["src", "index.ts"],
-  "exclude": ["src/**/__docs__", "src/**/__test__", "**/*.test.ts"]
+  "exclude": ["src/**/__docs__", "src/**/__test__"]
 }
